fixed temperature being integer instead of decimal

+++ models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, DateTime, ForeignKey, Integer, String
+from sqlalchemy import Column, DateTime, ForeignKey, Integer, String, DECIMAL
 from flask_sqlalchemy import SQLAlchemy
 from flask_wtf.csrf import CSRFProtect
 from flask_migrate import Migrate
@@ -27,7 +27,7 @@ class SensorData(db.Model):
     co2 = Column(Integer)
     humidity = Column(Integer)
     pressure = Column(Integer)
-    temperature = Column(Integer)
+    temperature = Column(DECIMAL(3,1))
     location = Column(Integer)
 
     def __str__(self):
